worker/requirements.txt
celery==5.3.6
redis==5.0.1
requests==2.31.0
python-dotenv==1.0.1

